requirements.txt
aiohttp==3.7.3
amqp==2.6.1
arrow==0.17.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
bcrypt==3.2.0
billiard==3.6.3.0
blessed==1.17.10
boto3==1.15.16
botocore==1.18.18
cached-property==1.5.2
celery==4.4.2
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.7.0rc1
dj-stripe==2.4.3
Django==3.1.7
django-bootstrap-modal-forms==2.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-extensions==2.2.5
django-hosts==4.0
django-ipware==3.0.2
django-sslserver==0.22
django-storages==1.10.1
django-taggit==1.2.0
django-timezone-field==4.1.1
django-tinymce==3.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
dnspython==2.1.0
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
flower==0.9.3
future==0.16.0
GDAL==3.1.3
geocoder==1.38.1
geographiclib==1.50
geoip2==4.1.0
geopy==2.0.0
geos==0.2.2
gunicorn==20.0.4
httpie==2.3.0
icalendar==4.0.7
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==3.2.0
kombu==4.6.11
lxml==4.5.2
MarkupSafe==1.1.1
maxminddb==2.0.3
multidict==5.1.0
oauthlib==3.1.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==7.0.0
postgis==1.0.4
proj==0.2.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pycountry==20.7.3
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==19.0.0
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
rabbitmq==0.2.0
ratelim==0.1.6
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.5
s3transfer==0.3.3
semantic-version==2.8.5
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
stripe==2.56.0
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.3
tornado==5.1.1
twilio==6.52.0
typing-extensions==3.7.4.3
urllib3==1.25.11
vine==1.3.0
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==0.16.0
whitenoise==5.2.0
yarl==1.6.3
